@@ -2,17 +2,21 @@
   "name": "summer_intensive",
   "version": "1.0.0",
   "dependencies": {
+    "@u-wave/react-vimeo": "^0.2.0",
     "@vimeo/player": "^2.1.0",
     "babel-plugin-add-module-exports": "^0.2.1",
     "babel-runtime": "^6.23.0",
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.17.2",
+    "btoa": "^1.1.2",
     "connect-history-api-fallback": "^1.3.0",
     "dotenv": "^4.0.0",
-    "electron-to-chromium": "^1.3.14",
+    "electron-to-chromium": "^1.3.15",
+    "enhanced-resolve": "^3.3.0",
     "express": "^4.15.3",
     "express-history-api-fallback": "^2.2.1",
     "express-jwt": "^5.3.0",
+    "halogen": "^0.2.0",
     "history": "^4.6.3",
     "isomorphic-fetch": "^2.2.1",
     "jquery": "^3.2.1",
@@ -23,40 +27,47 @@
     "num2fraction": "^1.2.2",
     "path": "^0.12.7",
     "path-to-regexp": "^1.7.0",
-    "pg-promise": "^5.9.0",
+    "pg-promise": "^6.3.5",
     "pinkie-promise": "^2.0.1",
     "prop-types": "^15.5.10",
-    "react": "^15.5.4",
-    "react-bootstrap": "^0.31.0",
+    "react-bootstrap": "^0.31.1",
     "react-icons": "^2.2.5",
+    "react-redux": "^5.0.6",
     "react-router": "^4.1.1",
     "react-router-dom": "^4.1.1",
     "react-router-link": "^2.0.2",
+    "react-router-redux": "^4.0.8",
     "react-router-relative-link": "^1.0.3",
-    "react-vimeo": "^1.0.0"
+    "react-vimeo": "^1.0.0",
+    "redux": "^3.7.2",
+    "tapable": "^0.2.6",
+    "watchpack": "^1.3.1",
+    "webpack-hot-middleware": "^2.18.2"
   },
   "devDependencies": {
     "babel": "^6.5.2",
     "babel-cli": "^6.24.1",
-    "babel-core": "^6.11.4",
-    "babel-loader": "^6.2.4",
+    "babel-core": "^6.25.0",
+    "babel-loader": "^7.1.1",
     "babel-plugin-transform-runtime": "^6.23.0",
     "babel-plugin-transform-strict-mode": "^6.11.3",
     "babel-preset-es2015": "^6.9.0",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-0": "^6.24.1",
-    "css-loader": "^0.23.1",
+    "css-loader": "^0.28.4",
     "dotenv": "^2.0.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.22.0",
-    "html-webpack-template": "^5.0.0",
-    "postcss-loader": "^1.1.1",
-    "react": "^15.3.0",
-    "react-dom": "^15.5.4",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.1"
+    "extract-text-webpack-plugin": "^2.1.2",
+    "file-loader": "^0.11.2",
+    "html-webpack-plugin": "^2.29.0",
+    "html-webpack-template": "^6.0.1",
+    "postcss-loader": "^2.0.6",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
+    "style-loader": "^0.18.2",
+    "url-loader": "^0.5.9",
+    "webpack": "^3.5.5",
+    "webpack-dev-middleware": "^1.12.0",
+    "webpack-dev-server": "^2.5.1"
   },
   "scripts": {
     "clean": "rm -rf dist && rm -rf node_modules",
