@@ -7,44 +7,46 @@ export default class VideoList extends Component{
   constructor(props){
     super(props)
     this.state = {
-      Videos: [],
+      videos: [],
       unwatched: [],
       loading: false,
+      viewable: false
     }
 
   }
 onLoad(data){
     console.log(data.duration)
 }
-   componentWillMount() {
-        let user = localStorage.getItem('ind')
-     this.setState({loading: true})
-    fetch(`/videos/${user}`)
-    .then(response => response.json())
-    .then(json => json)
-    .then(videos =>(console.log(videos, "videos list"),
-      this.setState({
-      videos,
-      loading: false
-    })
-      ))
-  }
+componentDidMount(){
+  console.log('showing these', this.props.showVids)
+}
+  //  componentWillMount() {
+  //       let user = localStorage.getItem('ind')
+  //    this.setState({loading: true})
+  //   fetch(`/videos/viewed/${user}/${this.props.showVids}`)
+  //   .then(response => response.json())
+  //   .then(json => json)
+  //   .then(videos =>
+  //     this.setState({
+  //     videos,
+  //     loading: false
+  //   })
+  //     )
+  // }
 
 
   render(){
-    let {  videos, loading } = this.state
-    console.log(videos, 'watched')
-
+    let {  videos, loading, viewable } = this.state
+    let {showVids} = this.props
     return(
     <div>
         {(loading)?
           <span>loading...</span>:
           null
         }
-      {(videos)?
-        videos.map(
+      {(showVids)?
+        showVids.map(
           function(video, index) {
-
             return(
             <Video  key={index}
                     video_id={video.video_id}
@@ -52,7 +54,8 @@ onLoad(data){
                     src={video.source}
                     watched={video.watched}
                     course={false}
-                    video_num={video.vimeo_id} />
+                    video_num={video.vimeo_id}
+                    viewable={viewable}/>
                     )
                      }):
               <span> Currently 0 Videos </span>
