@@ -2,7 +2,7 @@ const db = require('../lib/db');
 
 function getAnswers(req, res, next){
   console.log(req.params.student, 'student id')
-  db.any("SELECT DISTINCT users.username, user_quiz.answer, answer_sheet.question_id, answer_sheet.question, answer_sheet.answer_long, answer_sheet.answer_short FROM users, user_quiz JOIN answer_sheet on user_quiz.question_id = answer_sheet.question_id WHERE users.user_id=($1);", [req.params.student])
+  db.any("SELECT users.email, user_quiz.answer, answer_sheet.question_id, answer_sheet.question, answer_sheet.answer_long, answer_sheet.answer_short FROM users JOIN user_quiz on user_quiz.user_id = users.user_id JOIN answer_sheet on user_quiz.question_id = answer_sheet.question_id WHERE users.user_id=($1) ORDER BY question_id ASC;", [req.params.student])
    .then((answers) => {
     console.log(answers, 'these are the answers')
       res.answers = answers;
@@ -14,8 +14,9 @@ function getAnswers(req, res, next){
     // "admin/answers"
 }
 
+
 function getUsername(req, res, next){
-  db.one('select username from users where username=($1)', [req.body])
+  db.one('select email from users where email=($1);', [req.body])
   .then((username) => {
     console.log(username)
       res.username = username;
