@@ -5,7 +5,7 @@ userRouter.route('/signup/:admin')
   .post(createUser, (req, res) => res.json(res.user || {message: 'user succesfully signed up'}));
 
 userRouter.route('/login')
-  .post( login, (req, res) => res.json(res.user || {message: 'user succesfully logged in'}));
+  .post( login, (req, res) => res.json(res.user || res.status));
 
 userRouter.route('/user/:user')
   .post(create_relation, (req, res)=> res.json({message: "created user video relation table"}))
