@@ -1,7 +1,7 @@
 const videoRouter = require('express').Router()
 const { getUnwatched, course_vid, update_watched, getWatched } = require('../models/video.js')
 
-videoRouter.route('/:user')
+videoRouter.route('/viewed/:user/:videos')
   .get(getWatched, (req, res) => res.json(res.videos || []));
 
 videoRouter.route('/non/:user')
