@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from 'react';
+import React, { Component } from 'react';
+import { PropTypes} from 'prop-types';
 import fetch from 'isomorphic-fetch';
 import {withRouter} from "react-router-dom";
 
@@ -6,80 +7,68 @@ export default class Login extends Component {
   constructor(props){
     super(props)
     this.state = {
-      currentToken: 0
+      currentToken: 0,
+      correctPass: true,
+      status: "You've entered the wrong email or password"
     }
     this.submit = this.submit.bind(this)
   }
    submit(e) {
     e.preventDefault()
-    console.log('clicked')
     fetch('/user/login', {
       method: 'POST',
       headers: {
         'content-type': 'application/json',
       },
       body: JSON.stringify({
-        username: this.refs.username.value,
+        email: this.refs.email.value,
         password: this.refs.password.value
       })
     })
     .then(r => r.json())
     .then((user) => {
-      let data = user
-      if (data.token == undefined || data.token == null){
-        console.log('is null')
-        console.log('error at login')
-      }else {
-      console.log(user, 'is user')
+      console.log(user, 'user')
+      if (user.error == 'no email'){
+          this.setState({correctPass: false})
+          this.setState({status: "This email hasn't been assigned"})
+        }else if(user.error =='invalid password'){
+          this.setState({correctPass: false})
+          this.setState({status: "You've entered the wrong email or password"})
+        }
+        else{
+      this.setState({correctPass: true})
       localStorage.setItem('token', user.token)
-      localStorage.setItem('ind', user.username)
-      console.log(user.admin)
-      if (user.admin == true){
-        console.log('admin')
-        let val = user.admin
-        val.toString()
+      localStorage.setItem('ind', user.email)
       this.props.next.props.history.push("/dashboard")
-      }else if (user.admin == false){
-        let val = user.admin
-        val.toString()
-        this.props.next.props.history.push('/dashboard')
       }
-      }
-      }
-
-      )
+      })
 }
+
   render(){
-    const {username, password} = this.props
+    const {email, password} = this.props
+    const {correctPass, status} = this.state
     return(
           <div>
+          {(correctPass == true)? null : <h5 style={{color: 'red'}}>{status}</h5> }
+
             <form onSubmit={this.submit} >
-      <div className="row">
-            <div className="col-md-6">
-              <div className="form-group">
-                <input id="username" type="text" required ref="username" placeholder="Username" className="form-control" />
-              </div>
-            </div>
-            <div className="col-md-6">
-              <div className="form-group">
-                <input id="password" ref="password" type="password" required placeholder="Password" className="form-control" />
-              </div>
-            </div>
-          </div>
           <div className="form-group">
-            <input type="text" placeholder="Email address" className="form-control" />
-          </div>
- <a onClick={this.submit} className="btn btn-primary btn-block" style={{outline: 'none', cursor: 'inherit', outlineOffset: '-2px', borderRadius: '4px', fontSize: '14px', backgroundColor: 'rgb(22, 160, 133)'}} src="http://placehold.it/800x600">Login</a>
+            <input id="email" type="text" required ref="email" placeholder="Email" className="form-control" />
+            </div>
+          <div className="form-group">
+            <input id="password" ref="password" type="password" required placeholder="Password" className="form-control" />
+            </div>
+ <button onClick={this.submit} className="btn btn-primary btn-block" style={{outline: 'none', cursor: 'inherit', outlineOffset: '-2px', borderRadius: '4px', fontSize: '14px', backgroundColor: 'rgb(22, 160, 133)'}} src="http://placehold.it/800x600">Login</button>
             </form>
           </div>
     )
   }
 }
 Login.defaultProps = {
-  username: "imani",
+  email: "imani",
   password: "imani"
 }
 Login.propTypes = {
-  username: PropTypes.string.isRequired,
+  email: PropTypes.string.isRequired,
   password: PropTypes.string.isRequired
 }
