@@ -1,15 +1,17 @@
-import React, { PropTypes, Component } from 'react';
-
+import React, { Component } from 'react';
+import { PropTypes} from 'prop-types';
 export default class Quiz extends Component{
  constructor(props){
     super(props)
     this.state= {
       selectedOption: "",
       loading: false,
-      visible: true
+      visible: true,
+      textCapture: ''
     }
     this.submit = this.submit.bind(this)
     this.handleOptionChange = this.handleOptionChange.bind(this)
+    this.textCapture = this.textCapture.bind(this)
   }
   submit(e){
     e.preventDefault()
@@ -18,6 +20,8 @@ export default class Quiz extends Component{
     let q_id = this.props.question_id
     let user_answer = this.state.selectedOption
     let username = localStorage.getItem("ind")
+    let short_answer = this.state.textCapture
+    if (!short_answer){
     this.setState({loading: true,
                     visible: false})
     fetch("/questions/post_q"+"/"+q_id+"/"+username+"/"+user_answer, {
@@ -27,6 +31,7 @@ export default class Quiz extends Component{
       },
       body: JSON.stringify({
         answered: true,
+        short_answer: null
       })
     })
     .then(response => response.json())
@@ -35,21 +40,34 @@ export default class Quiz extends Component{
       console.log(returned)
       )
     console.log('You have selected:', this.state.selectedOption)
+  } else if (short_answer){
+    this.setState({loading: true,
+                    visible: false})
+    fetch("/questions/post_q"+"/"+q_id+"/"+username+"/short_answer", {
+      method: 'POST',
+      headers: {
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify({
+        answered: true,
+        short_answer: short_answer
+      })
+    })
+    .then(response => response.json())
+    .then(json => json)
+    .then(returned =>
+      console.log(returned)
+      )
+  }
   }
   handleOptionChange(e) {
-console.log(e.target)
-    console.log(e.target.className)
   this.setState({
     selectedOption: e.target.value
   });
 }
-handle(e){
-
-  // this.setState({
-  //   selectedOption: e.target.value
-  // });
-  // console.log(e.target.value)
-  // console.log(this.state.selectedOption)
+textCapture(e){
+this.setState({textCapture: e.target.value})
+console.log(this.state.textCapture)
 }
   render(){
       const {question, ans_short, question_id, video_id, choice1, choice2, choice3 } = this.props
@@ -63,7 +81,7 @@ handle(e){
            {question}
            </label>
            {(ans_short == null)? <div>
-                                <input type="text" name={question_id} />
+                                <input ref='short_answer' onChange={this.textCapture} type="text" name={question_id} />
                                     </div>
                                     :
                                     <div>
