@@ -1,4 +1,5 @@
-import React, { PropTypes, Component } from 'react';
+import React, { Component } from 'react';
+import { PropTypes} from 'prop-types';
 import Quiz from '../Quiz/Quiz.jsx';
 import { Link } from 'react-router-dom';
 export default class QuizList extends Component{
