@@ -10,47 +10,34 @@ export default class Video extends Component{
     this.state = {
       loading: false,
       duration: 0,
-      currentTime: 0
+      currentTime: 0,
+      courseVal: false,
     }
-    this.videoFrame = this.videoFrame.bind(this)
-    this.timeupdate = this.timeupdate.bind(this)
     this.play = this.play.bind(this)
+    this.end = this.end.bind(this)
+    this.onTimeUpdate = this.onTimeUpdate.bind(this)
   }
-  videoFrame(){
-    e.preventDefault()
-    console.log('clicked video frame')
-  }
-  loaded(e){
-    console.log(e)
-    console.log(e.target)
-  }
-  componentDidMount() {
- window.addEventListener("message", this.timeupdate);
-}
- componentWillUnmount() {
- window.removeEventListener("message", this.handleFrameTasks);
- }
- handleFrameTasks(e){
-  console.log(e)
-  // console.log(e.data)
 
-}
-timeupdate(e){
-  let duration
-  let currentTime
-  console.log(e.data)
-  console.log(duration)
-  if (e.data.event = "timeupdate"){
-    this.setState({
-  duration: e.data.data.duration,
-  currentTime: e.data.data.seconds
+
+play(e){
+console.log('the video is playing')
+let user = localStorage.getItem('ind')
+let vid = this.props.video_id
+fetch(`/badges/update/${vid}/${user}`,{
+  method: 'POST',
+  headers:{
+    'content-type': 'application/json',
+  }
 })
+}
 
-  }
+onTimeUpdate(e){
+  console.log(e)
+}
 
-  if (this.state.currentTime < this.state.duration && this.state.currentTime>this.state.duration-3){
-    console.log('ended')
-    let vid_id = this.props.video_id
+end(){
+  console.log('video ended')
+  let vid_id = this.props.video_id
        let user = localStorage.getItem('ind')
      fetch('/videos/update/'+vid_id, {
       method: 'POST',
@@ -63,43 +50,42 @@ timeupdate(e){
       })
     })
 
+fetch(`/badges/updatec/${vid_id}/${user}`,{
+  method: 'POST',
+  headers:{
+    'content-type': 'application/json',
   }
-
-
-}
-play(e){
-
-this.setState({
-  duration: e.data.data.duration
 })
-console.log('the video is playing')
-
+  console.log(this.props.course_val, this.props.part_val, 'bitch')
 }
+
   render(){
-    const {className, src, allowFullScreen, watched, video_id, course, video_num} = this.props
-    console.log(this.state.duration)
-    console.log(this.state.currentTime)
-    console.log(course, 'course')
+    const {className, src, allowFullScreen, watched, video_id, course, video_num, course_val, part_val} = this.props
     return(
     <div className={className} >
-      {(course==true) ?<Vimeo onPlay={this.play}
-            className="vimeo_vid"
-            videoId={ video_num } />
+      {(course==true) ?
+        <Vimeo
+            videoId={ video_num }
+            onPlay={this.play}
+            className="vimeo_vid_course"
+            onTimeupdate={this.onTimeUpdate}
+            onEnded={this.end} />
             :
       <div className="grid col-md-4 col-sm-6 col-xs-12">
           <figure className="effect-lily">
-            <Vimeo onPlay={this.play}
-            className="vimeo_vid"
-            videoId={ video_num } />
+            <Vimeo
+            videoId={ video_num }
+            className="vimeo_vid" />
+
             <figcaption>
-              <div className="editContent" style={{outline: 'none', cursor: 'inherit'}}>
+              <div className="editContent" style={{outline: 'none', textShadow:'1px 1px #000000',  cursor: 'inherit'}}>
                 <h2>Item <strong>3</strong></h2>
               </div>
-              <div className="editContent" style={{outline: 'none', cursor: 'inherit'}}>
+              <div className="editContent" style={{outline: 'none', textShadow:'1px 1px #000000', cursor: 'inherit'}}>
                 <p>Beautifully subtle animated hover effect for your gallery</p>
               </div>
               {((watched==false && course==true) ||(watched==true))? null : <Link to={`/courses/${video_id}`}>Quiz</Link>}
-              <a href="#"></a>
+              <a href={`video/${video_id}`}></a>
             </figcaption>
           </figure>
         </div>
