@@ -30,7 +30,7 @@ export default class AdminAnswers extends Component {
           <div className="editContent">
             <h2>Question One</h2>
           </div>
-          <div className="editContent" style={{outline: 'rgba(233, 94, 94, 0.5) solid 2px', outlineOffset: '-2px', cursor: 'pointer'}} contenteditable="true" spellcheck="true" data-medium-editor-element="true" role="textbox" aria-multiline="true" medium-editor-index="0" data-placeholder="Type your text" data-medium-focused="true">
+          <div className="editContent" style={{outline: 'rgba(233, 94, 94, 0.5) solid 2px', outlineOffset: '-2px', cursor: 'pointer'}} data-medium-editor-element="true" role="textbox" aria-multiline="true" data-placeholder="Type your text" data-medium-focused="true">
             <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam massa ex, suscipit id ligula at, luctus tempus dolor. Etiam sodales posuere ligula, eu aliquet nunc commodo vitae. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos.</p>
             <p>Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nam massa ex, suscipit id ligula at, luctus tempus dolor. Etiam sodales posuere ligula, eu aliquet nunc commodo vitae. Class aptent taciti sociosqu ad litora torquent per conubia nostra, per inceptos himenaeos.</p>
           </div>
@@ -96,6 +96,8 @@ export default class AdminAnswers extends Component {
     </div>
   </section>
       <UserAnswersList user={this.state.student}/>
+      <label>User Assesment</label>
+      <input type="text" ref='user_assesment' />
       </div>
       )
   }
