@@ -8,28 +8,40 @@ const BUILD_DIR         = path.resolve(__dirname, 'dist');
 const APP_DIR           = path.resolve(__dirname, 'src');
 
 module.exports = {
-  entry: `${APP_DIR}/index.js`,
+  entry: [
+  'webpack-hot-middleware/client',
+  `${APP_DIR}/index.js`,
+  ],
   output: {
-    path: BUILD_DIR,
-    filename: '/js/[name].js',
+    path: `${BUILD_DIR}`,
+    publicPath: `${BUILD_DIR}`,
+    filename: `./js/[name].js`,
   },
   cache: true,
-  debug: true,
   devtool: 'eval-source-map',
   stats: {
     colors: true,
     reasons: true
   },
   plugins: [
+    new webpack.NoEmitOnErrorsPlugin(),
+    new webpack.optimize.OccurrenceOrderPlugin(),
+    new webpack.HotModuleReplacementPlugin(),
     new HtmlWebpackPlugin({
       title: 'Urban Arts Summer Intensive',
       xhtml: true,
       inject: false,
-      template: require('html-webpack-template'),
+      debug: true,
+      template: 'src/index.html',
       appMountId: 'root-container'
     }),
-    new ExtractTextPlugin('/css/[name].css', {
+    new ExtractTextPlugin({
+      filename:'/css/[name].css',
+      disable: false,
       allChunks: true
+    }),
+    new webpack.LoaderOptionsPlugin({
+    debug: true
     })
   ],
   //just added this
@@ -38,31 +50,42 @@ module.exports = {
   },
 
   module : {
-    include: path.join(__dirname, 'src'),
-    loaders: [
+    rules: [
       {
         test: /\.css$/,
-        loader: ExtractTextPlugin.extract('style-loader', 'css-loader')
+        use: ExtractTextPlugin.extract({
+          fallback: "style-loader",
+          use: ["css-loader"]
+  })
+
+
       },
       {
         test: /\.svg$/,
-        loader: 'file-loader?name=/img/[name].[hash:base64:5].[ext]'
+        include: path.join(__dirname, 'src'),
+        loader: 'file-loader?name=/img/[name].[hash:base64:5].[ext'
       },
       {
         test: /\.gif$/,
+        include: path.join(__dirname, 'src'),
         loader: 'file-loader?name=/img/[name].[hash:base64:5].[ext]'
       },
       { test: /\.jpg$/,
+        include: path.join(__dirname, 'src'),
         loader: 'file-loader?name=/img/[name].[hash:base64:5].[ext]'
       },
       {
         test: /\.(js|jsx)$/,
-        loader: 'babel'
+        exclude: [path.resolve(__dirname, "node_modules")],
+        loader: `babel-loader`
       },
       {
         test: /\.otf(\?v=\d+\.\d+\.\d+)?$/,
+        include: path.join(__dirname, 'src'),
         loader: 'file-loader?name=/fonts/[name].[ext]'
+
       }
     ]
   }
+
 };
